Refactor methods into Schema, OleDbSchema Classes
