+++ DatabaseLibraryNet/OleDbSchema.cs
@@ -11,6 +11,7 @@ namespace DigitalZenWorks.Database.ToolKit
 	using System.Data.OleDb;
 	using System.Globalization;
 	using System.Runtime.Versioning;
+	using global::Common.Logging;
 
 	/// Class <c>OleDbSchema.</c>
 	/// <summary>
@@ -21,6 +22,9 @@ namespace DigitalZenWorks.Database.ToolKit
 #endif
 	public class OleDbSchema : IDisposable
 	{
+		private static readonly ILog Log = LogManager.GetLogger(
+			System.Reflection.MethodBase.GetCurrentMethod().DeclaringType);
+
 		/// <summary>
 		/// Represents an OleDb open connection to a data source.
 		/// </summary>
@@ -79,6 +83,88 @@ namespace DigitalZenWorks.Database.ToolKit
 
 				return schemaTable;
 			}
+		}
+
+		/// <summary>
+		/// Creates a new Column instance by extracting and formatting column
+		/// metadata from the specified DataRow.
+		/// </summary>
+		/// <remarks>The returned Column reflects the schema information
+		/// present in the DataRow, including type, length, nullability,
+		/// default value, and position. The method expects the DataRow to
+		/// follow a specific schema layout, such as that returned from
+		/// database schema queries.</remarks>
+		/// <param name="row">The DataRow containing column metadata. Must
+		/// include fields such as COLUMN_NAME, DATA_TYPE, and other relevant
+		/// schema information.</param>
+		/// <returns>A Column object populated with properties derived from
+		/// the values in the provided DataRow.</returns>
+		public static Column FormatColumnFromDataRow(DataRow row)
+		{
+			Column column = null;
+
+			if (row != null)
+			{
+				column = new ();
+				column.Name = row["COLUMN_NAME"].ToString();
+
+				switch ((int)row["DATA_TYPE"])
+				{
+					case 3: // Number
+						column.ColumnType = ColumnType.Number;
+						break;
+					case 130: // String
+						string flags = row["COLUMN_FLAGS"].ToString();
+
+						if (int.Parse(flags, CultureInfo.InvariantCulture) > 127)
+						{
+							column.ColumnType = ColumnType.Memo;
+						}
+						else
+						{
+							column.ColumnType = ColumnType.String;
+						}
+
+						break;
+					case 7: // Date
+						column.ColumnType = ColumnType.DateTime;
+						break;
+					case 6: // Currency
+						column.ColumnType = ColumnType.Currency;
+						break;
+					case 11: // Yes/No
+						column.ColumnType = ColumnType.YesNo;
+						break;
+					case 128: // OLE
+						column.ColumnType = ColumnType.Ole;
+						break;
+				}
+
+				if (!row.IsNull("CHARACTER_MAXIMUM_LENGTH"))
+				{
+					string maxLength =
+						row["CHARACTER_MAXIMUM_LENGTH"].ToString();
+
+					column.Length =
+						int.Parse(maxLength, CultureInfo.InvariantCulture);
+				}
+
+				if (row["IS_NULLABLE"].ToString() == "True")
+				{
+					column.Nullable = true;
+				}
+
+				if (row["COLUMN_HASDEFAULT"].ToString() == "True")
+				{
+					column.DefaultValue = row["COLUMN_DEFAULT"].ToString();
+				}
+
+				string position = row["ORDINAL_POSITION"].ToString();
+				column.Position =
+					int.Parse(position, CultureInfo.InvariantCulture);
+			}
+
+			return column;
 		}
 
 		/// <summary>
@@ -148,6 +234,34 @@ namespace DigitalZenWorks.Database.ToolKit
 		}
 
 		/// <summary>
+		/// Retrieves a table definition, including its columns, for the
+		/// specified table name.
+		/// </summary>
+		/// <remarks>The returned <see cref="Table"/> includes all columns as
+		/// defined in the data source. If the table does not exist, the
+		/// result may be an empty table definition.</remarks>
+		/// <param name="tableName">The name of the table to retrieve. Cannot
+		/// be null or empty.</param>
+		/// <returns>A <see cref="Table"/> object representing the specified
+		/// table and its columns.</returns>
+		public Table GetTable(string tableName)
+		{
+			Table table = new(tableName);
+
+			Log.Info("Getting Columns for " + tableName);
+			DataTable dataColumns = GetTableColumns(tableName);
+
+			foreach (DataRow dataColumn in dataColumns.Rows)
+			{
+				Column column = FormatColumnFromDataRow(dataColumn);
+
+				table.AddColumn(column);
+			}
+
+			return table;
+		}
+
+		/// <summary>
 		/// Gets the column names from the given table.
 		/// </summary>
 		/// <param name="tableName">The name of the table.</param>
@@ -186,6 +300,55 @@ namespace DigitalZenWorks.Database.ToolKit
 		}
 
 		/// <summary>
+		/// Sets the primary key for the specified table based on the
+		/// information provided in the given data row.
+		/// </summary>
+		/// <remarks>If the primary key column is of integer type, its type is
+		/// set to AutoNumber. This method currently supports only single
+		/// -column primary keys; composite primary keys are
+		/// not handled.</remarks>
+		/// <param name="row">The data row containing table metadata,
+		/// including the table name and primary key information. Must not be
+		/// null and must contain valid 'TABLE_NAME' and 'COLUMN_NAME'
+		/// fields.</param>
+		/// <returns>A Table object with its PrimaryKey property set according
+		/// to the primary key defined in the data row.</returns>
+		public Table SetPrimaryKey(DataRow row)
+		{
+			Table table;
+
+			if (row == null)
+			{
+				throw new ArgumentNullException(nameof(row));
+			}
+			else
+			{
+				object nameRaw = row["TABLE_NAME"];
+				string tableName = nameRaw.ToString();
+
+				table = GetTable(tableName);
+
+				DataTable primaryKeys = GetPrimaryKeys(tableName);
+
+				// TODO: This assumes only a single primary key.  Need to
+				// compensate for composite primary keys.
+				DataRow primaryKeyRow = primaryKeys.Rows[0];
+				nameRaw = primaryKeyRow["COLUMN_NAME"];
+				table.PrimaryKey = nameRaw.ToString();
+
+				// If PK is an integer change type to AutoNumber
+				Column primaryKey = SetPrimaryKeyType(table);
+
+				if (primaryKey != null)
+				{
+					table.Columns[table.PrimaryKey] = primaryKey;
+				}
+			}
+
+			return table;
+		}
+
+		/// <summary>
 		/// Dispose.
 		/// </summary>
 		/// <param name="disposing">Indicates whether the object is
@@ -197,6 +360,25 @@ namespace DigitalZenWorks.Database.ToolKit
 				oleDbConnection?.Close();
 				oleDbConnection = null;
 			}
+		}
+
+		// If primary key is an integer, change type to AutoNumber.
+		private static Column SetPrimaryKeyType(Table table)
+		{
+			Column primaryKey = null;
+
+			string primaryKeyName = table.PrimaryKey;
+			if (!string.IsNullOrWhiteSpace(primaryKeyName))
+			{
+				primaryKey = table.Columns[primaryKeyName];
+
+				if (primaryKey.ColumnType == ColumnType.Number)
+				{
+					primaryKey.ColumnType = ColumnType.AutoNumber;
+				}
+			}
+
+			return primaryKey;
 		}
 	}
 }
